+++ src/main/groovy/ru/vyarus/gradle/plugin/lib/JavaLibPlugin.groovy
@@ -6,32 +6,25 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.artifacts.Configuration
-import org.gradle.api.artifacts.PublishArtifact
-import org.gradle.api.internal.artifacts.dsl.LazyPublishArtifact
 import org.gradle.api.java.archives.Attributes
-import org.gradle.api.plugins.BasePlugin
-import org.gradle.api.plugins.GroovyPlugin
-import org.gradle.api.plugins.JavaPlugin
-import org.gradle.api.plugins.JavaPluginConvention
-import org.gradle.api.plugins.JavaPluginExtension
-import org.gradle.api.plugins.ProjectReportsPlugin
+import org.gradle.api.plugins.*
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
 import org.gradle.api.publish.maven.tasks.PublishToMavenLocal
 import org.gradle.api.publish.maven.tasks.PublishToMavenRepository
 import org.gradle.api.publish.tasks.GenerateModuleMetadata
 import org.gradle.api.reporting.ConfigurableReport
-import org.gradle.api.reporting.Report
 import org.gradle.api.reporting.SingleFileReport
+import org.gradle.api.reporting.dependencies.HtmlDependencyReportTask
 import org.gradle.api.tasks.Copy
-import org.gradle.api.tasks.TaskProvider
-import org.gradle.api.tasks.bundling.Jar
+import org.gradle.api.tasks.SourceSetContainer
 import org.gradle.api.tasks.compile.GroovyCompile
 import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.api.tasks.javadoc.Javadoc
 import org.gradle.api.tasks.testing.Test
 import org.gradle.api.tasks.testing.TestReport
+import org.gradle.jvm.tasks.Jar
+import org.gradle.plugins.signing.SigningExtension
 import org.gradle.plugins.signing.SigningPlugin
 import org.gradle.process.internal.JvmOptions
 import org.gradle.testing.jacoco.plugins.JacocoPlugin
@@ -45,10 +38,9 @@ import java.nio.charset.StandardCharsets
  * Plugin performs common configuration for java or groovy library:
  * <ul>
  *     <li> Fills manifest for main jar, add pom and generated pom.properties inside jar (as maven do)
- *     <li> Add sourcesJar task
- *     <li> Add javadocJar or (and) groovydocJar tasks
- *     <li> Configures maven publication named 'maven' with all jars (jar, sources, javadoc and maybe groovydoc)
- *     <li> Applies 'ru.vyarus.pom' plugin which fixes pom dependencies, adds 'pom' closure fot simpler pom
+ *     <li> Add sourcesJar and javadocJar tasks (gradle native way)
+ *     <li> Configures maven publication named 'maven' (with all jars)
+ *     <li> Applies 'ru.vyarus.pom' plugin which fixes pom dependencies, adds 'pom' closure for simpler pom
  *     configuration
  *     <li> Configure signing if "signing" plugin applied
  *     <li> Add 'install' task as shortcut for publishToMavenLocal
@@ -74,8 +66,7 @@ import java.nio.charset.StandardCharsets
  * <p>
  * In case of gradle plugin (java-gradle-plugin + plugin-publish) "pluginMaven" publication will be created for
  * plugins portal publication, but java-lib plugin will still use "maven" publication (for maven central
- * publication). Plugin-publish also creates it's own javadoc and sources tasks, so manually registering
- * artifacts (only) to prevent this.
+ * publication).
  * Overall, in case of gradle plugin, 2 maven publications should be used, but exactly the same artifacts would
  * be published everywhere (plugin portal will additionally receive alias publications).
  *
@@ -87,19 +78,13 @@ import java.nio.charset.StandardCharsets
 @CompileStatic(TypeCheckingMode.SKIP)
 class JavaLibPlugin implements Plugin<Project> {
 
-    private static final String BUILD_GROUP = 'build'
-    private static final String JAVADOC_JAR = 'javadocJar'
-    private static final String GROOVYDOC_JAR = 'groovydocJar'
-
     @Override
-    @SuppressWarnings('MethodSize')
     void apply(Project project) {
         // always creating extension to avoid hard to track mis-references in multi-module projects
         JavaLibExtension extension = project.extensions.create('javaLib', JavaLibExtension, project)
 
         // partial activation for java-platform plugin (when root module is a BOM)
-        // not by type because plugin was added in gradle 5.2
-        project.plugins.withId('java-platform') {
+        project.plugins.withType(JavaPlatformPlugin) {
             project.plugins.apply(PomPlugin)
             // different name used for publication
             MavenPublication bom = configureBomPublication(project)
@@ -119,14 +104,7 @@ class JavaLibPlugin implements Plugin<Project> {
             MavenPublication publication = configureMavenPublication(project)
             configureEncoding(project)
             configureJar(project, publication)
-            if (GradleVersion.current() < GradleVersion.version('7.6')) {
-                addSourcesJarTask(project, publication, extension)
-                addJavadocJarTask(project, publication, extension)
-                addGroovydocJarTask(project, publication, extension)
-            } else {
-                // since gradle 7.6 - use native integration
-                addJavadocAndSourcesNative(project, extension)
-            }
+            addJavadocAndSourceJars(project, extension)
             configureGradleMetadata(project, extension)
             disablePublication(project, extension)
             configureSigning(project, publication)
@@ -200,7 +178,7 @@ class JavaLibPlugin implements Plugin<Project> {
                         'groupId': "${ -> project.group }",
                         'artifactId': "${ -> project.name }",
                 ])
-                outputs.file "$project.buildDir/generatePomPropertiesFile/pom.properties"
+                outputs.file project.layout.buildDirectory.file('generatePomPropertiesFile/pom.properties')
                 doLast {
                     File file = outputs.files.singleFile
                     file.parentFile.mkdirs()
@@ -220,93 +198,32 @@ class JavaLibPlugin implements Plugin<Project> {
                     putIfAbsent(attributes, 'Built-Date', new Date())
                     putIfAbsent(attributes, 'Built-JDK', System.getProperty('java.version'))
                     putIfAbsent(attributes, 'Built-Gradle', project.gradle.gradleVersion)
-                    putIfAbsent(attributes, 'Target-JDK', project.targetCompatibility)
+                    putIfAbsent(attributes, 'Target-JDK', project.extensions
+                            .getByType(JavaPluginExtension).targetCompatibility)
                 }
             }
 
             project.tasks.named('processResources', Copy).configure {
                 it.with {
                     into("META-INF/maven/$project.group/$project.name") {
-                        from project.tasks.getByName("generatePomFileFor${publication.name.capitalize()}Publication")
+                        from project.tasks.named("generatePomFileFor${publication.name.capitalize()}Publication")
                         rename '.*.xml', 'pom.xml'
-                        from project.tasks.getByName('generatePomPropertiesFile')
+                        from project.tasks.named('generatePomPropertiesFile')
                     }
-                }
-            }
-        }
-    }
-
-    private void addSourcesJarTask(Project project, MavenPublication publication, JavaLibExtension extension) {
-        TaskProvider<Jar> sourcesJar = project.tasks.register('sourcesJar', Jar) {
-            it.with {
-                dependsOn project.tasks.named('classes')
-                group = BUILD_GROUP
-                description = 'Assembles a jar archive containing the main sources'
-                from project.sourceSets.main.allSource
-                archiveClassifier.set('sources')
-            }
-        }
-        registerArtifact(project, publication, sourcesJar)
-        project.afterEvaluate {
-            sourcesJar.configure { enabled = extension.addSources }
-        }
-    }
-
-    private void addJavadocJarTask(Project project, MavenPublication publication, JavaLibExtension extension) {
-        // apply only if java sources exist
-        boolean hasJavaSources = project.sourceSets.main.java.srcDirs.find { it.exists() }
-        if (hasJavaSources) {
-            TaskProvider<Jar> javadocJar = project.tasks.register(JAVADOC_JAR, Jar) {
-                it.with {
-                    dependsOn project.tasks.javadoc
-                    group = BUILD_GROUP
-                    description = 'Assembles a javadoc jar'
-                    archiveClassifier.set('javadoc')
-                    // configuration is delayed so it is ok to reference task instance here
-                    from project.tasks.javadoc.destinationDir
-                }
-            }
-            registerArtifact(project, publication, javadocJar)
-            project.afterEvaluate {
-                javadocJar.configure { enabled = extension.addJavadoc }
-            }
-        }
-    }
-
-    private void addGroovydocJarTask(Project project, MavenPublication publication, JavaLibExtension extension) {
-        // apply only if groovy enabled
-        project.plugins.withType(GroovyPlugin) {
-            // apply only if groovy sources exist
-            boolean hasGroovySources = project.sourceSets.main.groovy.srcDirs.find { it.exists() }
-            if (hasGroovySources) {
-                TaskProvider<Jar> groovydocJar = project.tasks.register(GROOVYDOC_JAR, Jar) {
-                    it.with {
-                        dependsOn project.tasks.groovydoc
-                        group = BUILD_GROUP
-                        description = 'Assembles a groovydoc jar'
-                        // very important to have at least one javadoc package, because otherwise maven cantral
-                        // would not accept package
-                        archiveClassifier.set(project.tasks.findByName(JAVADOC_JAR) ? 'groovydoc' : 'javadoc')
-                        from project.tasks.groovydoc.destinationDir
-                    }
-                }
-                registerArtifact(project, publication, groovydocJar)
-                project.afterEvaluate {
-                    groovydocJar.configure { enabled = extension.addJavadoc }
                 }
             }
         }
     }
 
     @SuppressWarnings('Indentation')
-    private void addJavadocAndSourcesNative(Project project, JavaLibExtension extension) {
+    private void addJavadocAndSourceJars(Project project, JavaLibExtension extension) {
         // use gradle native configuration method to aovid clashes with plugin-publish 1.0
         JavaPluginExtension javaExt = project.extensions.getByType(JavaPluginExtension)
         javaExt.withJavadocJar()
         javaExt.withSourcesJar()
         project.afterEvaluate {
             if (!extension.addJavadoc || !extension.addSources) {
-                project.convention.getPlugin(JavaPluginConvention).sourceSets
+                project.extensions.getByType(JavaPluginExtension).sourceSets
                         .named('main').configure { sourceSet ->
                     if (!extension.addJavadoc) {
                         project.tasks.named(sourceSet.javadocJarTaskName).configure {
@@ -324,14 +241,13 @@ class JavaLibPlugin implements Plugin<Project> {
     }
 
     private MavenPublication configureMavenPublication(Project project) {
+        // Configure publication:
         // java-gradle-plugin will create its own publication pluginMaven, but still plugin configures separate
         // maven publication because java-gradle-plugin most likely will be applied after java-lib and so
         // it's not possible to detect it for sure. But it's not a problem: pom will be corrected for both
-        // (and in any case portal does not use this pom). More importantly, to prevent plugin-publish to create
-        // its own javadoc and sources tasks (its done in later in artifacts method)
+        // (and in any case portal does not use this pom).
         // NOTE: java-gradle-plugin will also create alias publications for each plugin, but we will simply don't
-        // use them during bintray publication
-        // configure publication
+        // use them
         MavenPublication publication = project.extensions
                 .findByType(PublishingExtension)
                 .publications
@@ -394,8 +310,9 @@ class JavaLibPlugin implements Plugin<Project> {
     private void configureSigning(Project project, MavenPublication publication) {
         project.plugins.withType(SigningPlugin) {
             // https://docs.gradle.org/current/userguide/signing_plugin.html#sec:signatory_credentials
-            project.signing.sign publication
-            project.signing.required = { !project.version.toString().endsWith('SNAPSHOT') }
+            SigningExtension ext = project.extensions.getByType(SigningExtension)
+            ext.sign publication
+            ext.required = { !project.version.toString().endsWith('SNAPSHOT') }
         }
     }
 
@@ -403,7 +320,7 @@ class JavaLibPlugin implements Plugin<Project> {
         project.afterEvaluate {
             if (extension.autoModuleName) {
                 // java 11 auto module name
-                project.tasks.jar.manifest {
+                (project.tasks.jar as Jar).manifest {
                     attributes 'Automatic-Module-Name': extension.autoModuleName
                 }
             }
@@ -435,9 +352,8 @@ class JavaLibPlugin implements Plugin<Project> {
     private void enableJacocoXmlReport(Project project) {
         // by default jacoco xml report is disabled, but its required for coverage services
         project.plugins.withType(JacocoPlugin) {
-            JacocoReport task = project.tasks.findByName('jacocoTestReport') as JacocoReport
-            if (task) {
-                enableReport(task.reports.xml)
+            project.tasks.named('jacocoTestReport').configure {
+                (it as JacocoReport).reports.xml.required.set(true)
             }
         }
     }
@@ -489,46 +405,48 @@ class JavaLibPlugin implements Plugin<Project> {
                         // show task in common place
                         group = 'verification'
                         executionData project.files(projectsWithCoverage
-                                .collect { it.file("${it.buildDir}/jacoco/test.exec") })
+                                .collect { it.layout.buildDirectory.file('jacoco/test.exec') })
                                 .filter { it.exists() }
-                        sourceDirectories.from = project.files(projectsWithCoverage.sourceSets.main.allSource.srcDirs)
-                        classDirectories.from = project.files(projectsWithCoverage.sourceSets.main.output)
+                        sourceDirectories.from = selectFiles(projectsWithCoverage) { SourceSetContainer sourceSets ->
+                            sourceSets.main.allSource.srcDirs
+                        }
+                        classDirectories.from = selectFiles(projectsWithCoverage) { SourceSetContainer sourceSets ->
+                            sourceSets.main.output.files
+                        }
                         // use same location as in single-module case
                         reportDestination(reports.xml, project, 'reports/jacoco/test/jacocoTestReport.xml')
                         reportDestination(reports.html, project, 'reports/jacoco/test/html/')
-                        enableReport(reports.xml)
+                        reports.xml.required.set(true)
                     }
                 }
             }
 
             // aggregated html dependency report
             project.plugins.withType(ProjectReportsPlugin) {
-                project.htmlDependencyReport.projects = project.allprojects
+                project.tasks.named('htmlDependencyReport').configure {
+                    (it as HtmlDependencyReportTask).projects = project.allprojects
+                }
             }
         }
+    }
+
+    private Set<File> selectFiles(Set<Project> projects, Closure<Set<File>> extractor) {
+        Set<File> res = []
+        projects.forEach {
+            res.addAll(
+                    extractor.call(
+                            GradleVersion.current() < GradleVersion.version('7.6')
+                                    ? it.sourceSets
+                                    : it.extensions.getByType(JavaPluginExtension).sourceSets // added in 7.1
+                    )
+            )
+        }
+        res
     }
 
     private void putIfAbsent(Attributes attributes, String name, Object value) {
         if (!attributes.containsKey(name)) {
             attributes.put(name, value)
-        }
-    }
-
-    private void registerArtifact(Project project, MavenPublication publication, TaskProvider task) {
-        // https://github.com/gradle/gradle/issues/6246
-        PublishArtifact artifact = new LazyPublishArtifact(task)
-
-        // maven publication registration
-        publication.artifact artifact
-
-        // plugin-publish create its own sources and javadoc tasks, but only if
-        // archives configuration contains anything except main jar, so register custom tasks directly to avoid
-        // duplicate artifacts (see PublishPlugin implementation)
-        if (project.plugins.hasPlugin('com.gradle.plugin-publish')) {
-            Configuration archives = project.configurations.findByName('archives')
-            if (archives && !archives.artifacts.contains(artifact)) {
-                project.artifacts.add('archives', artifact)
-            }
         }
     }
 
@@ -542,14 +460,6 @@ class JavaLibPlugin implements Plugin<Project> {
             } else {
                 report.outputLocation.set(project.layout.buildDirectory.dir(path))
             }
-        }
-    }
-
-    private void enableReport(Report report) {
-        if (GradleVersion.current() < GradleVersion.version('7.0')) {
-            report.enabled = true
-        } else {
-            report.required.set(true)
         }
     }
 }