+++ src/test/groovy/ru/vyarus/gradle/plugin/lib/LegacyModeKitTest.groovy
@@ -102,7 +102,7 @@ class LegacyModeKitTest extends AbstractKitTest {
         deploy.exists()
         def baseName = artifactId + '-1.0'
         withoutModuleFile(deploy) ==
-                ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar", "${baseName}-groovydoc.jar"] as Set<String>
+                ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar"] as Set<String>
     }
 
     def "Check install for no sources"() {
@@ -132,7 +132,7 @@ class LegacyModeKitTest extends AbstractKitTest {
         deploy.exists()
         def baseName = artifactId + '-1.0'
         withoutModuleFile(deploy) ==
-                ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar"] as Set<String>
+                ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar"] as Set<String>
     }
 
     def "Check behaviour on test sources"() {
@@ -164,7 +164,7 @@ class LegacyModeKitTest extends AbstractKitTest {
         deploy.exists()
         def baseName = artifactId + '-1.0'
         withoutModuleFile(deploy) as Set ==
-                ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar"] as Set<String>
+                ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar"] as Set<String>
     }
 
     def "Check publication override"() {
@@ -178,6 +178,11 @@ class LegacyModeKitTest extends AbstractKitTest {
 
             group 'ru.vyarus'
             version 1.0
+            
+            javaLib {
+                // artifacts could be changed when metadata enabled
+                withoutGradleMetadata()
+            }
 
             publishing.publications.maven.artifacts = [jar, javadocJar]
         """